# src/main/resources/openapi/NotificationsAPIs.yaml
openapi: 3.0.0
info:
  title: Notifications APIs
  version: '1.15.1'
  description: Notifications APIs
  contact:
    name: Krister Bone
    email: krister.bone@valtech.com
#
# --------------------------------------------------------------------------------
#
paths:
...
  '/templates/photo-resubmission/preview':
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Generate Photo Resubmission Template Preview
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    post:
      summary: Generates a text and html preview for a photo resubmission template type and personalisation data
      description: Generates a text and html preview for a photo resubmission template type and personalisation data
      operationId: generate-photo-resubmission-template-preview
      tags:
        - Generate Photo Resubmission Template Preview
      requestBody:
        $ref: '#/components/requestBodies/GeneratePhotoResubmissionTemplatePreview'
      responses:
        '201':
          $ref: '#/components/responses/GenerateTemplatePreview'
        '400':
          $ref: '#/components/responses/400ErrorResponse'
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/templates/photo-resubmission/preview'
        requestParameters:
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: POST

  #
  # Generate ID Document Resubmission Template Preview
  # --------------------------------------------------------------------------------
  '/templates/id-document-resubmission/preview':
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Generate ID Document Resubmission Template Preview
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    post:
      summary: Generates a text and html preview for an identity document resubmission template type and personalisation data
      description: Generates a text and html preview for an identity document resubmission template type and personalisation data
      operationId: generate-id-document-resubmission-template-preview
      tags:
        - Generate ID Document Resubmission Template Preview
      requestBody:
        $ref: '#/components/requestBodies/GenerateIdDocumentResubmissionTemplatePreview'
      responses:
        '201':
          $ref: '#/components/responses/GenerateTemplatePreview'
        '400':
          $ref: '#/components/responses/400ErrorResponse'
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/templates/id-document-resubmission/preview'
        requestParameters:
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: POST

  #
  # Generate ID Document Required Template Preview
  # --------------------------------------------------------------------------------
  '/templates/id-document-required/preview':
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Generate ID Document Required Template Preview
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    post:
      summary: Generates a text and html preview for an ID document required template type and personalisation data
      description: Generates a text and html preview for an ID document required template type and personalisation data
      operationId: generate-id-document-required-template-preview
      tags:
        - Generate ID Document Required Template Preview
      requestBody:
        $ref: '#/components/requestBodies/GenerateIdDocumentRequiredTemplatePreview'
      responses:
        '201':
          $ref: '#/components/responses/GenerateTemplatePreview'
        '400':
          $ref: '#/components/responses/400ErrorResponse'
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/templates/id-document-required/preview'
        requestParameters:
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: POST

  #
  # Generate Application Rejected Template Preview
  # --------------------------------------------------------------------------------
  '/templates/application-rejected/preview':
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Generate Application Rejected Template Preview
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    post:
      summary: Generates a text preview for an application rejected template type and personalisation data
      description: Generates a text preview for an application rejected template type and personalisation data
      operationId: generate-application-rejected-template-preview
      tags:
        - Generate Application Rejected Template Preview
      requestBody:
        $ref: '#/components/requestBodies/GenerateApplicationRejectedTemplatePreview'
      responses:
        '201':
          $ref: '#/components/responses/GenerateTemplatePreview'
        '400':
          $ref: '#/components/responses/400ErrorResponse'
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/templates/application-rejected/preview'
        requestParameters:
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: POST

  #
  # Generate Application Approval Template Preview
  # --------------------------------------------------------------------------------
  '/templates/application-approved/preview':
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Generate Application Approved Template Preview
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    post:
      summary: Generates a text preview for an application approved template type and personalisation data
      description: Generates a text preview for an application approved template type and personalisation data
      operationId: generate-application-approved-template-preview
      tags:
        - Generate Application Approved Template Preview
      requestBody:
        $ref: '#/components/requestBodies/GenerateApplicationApprovedTemplatePreview'
      responses:
        '201':
          $ref: '#/components/responses/GenerateTemplatePreview'
        '400':
          $ref: '#/components/responses/400ErrorResponse'
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/templates/application-approved/preview'
        requestParameters:
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: POST


  #
  # Generate Rejected Document Template Preview
  # --------------------------------------------------------------------------------
  '/templates/rejected-document/preview':
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Generate Rejected Document Template Preview
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    post:
      summary: Generates a text preview for a rejected document template type and personalisation data
      description: Generates a text preview for a rejected document template type and personalisation data
      operationId: generate-rejected-document-template-preview
      tags:
        - Generate Rejected Document Template Preview
      requestBody:
        $ref: '#/components/requestBodies/GenerateRejectedDocumentTemplatePreview'
      responses:
        '201':
          $ref: '#/components/responses/GenerateTemplatePreview'
        '400':
          $ref: '#/components/responses/400ErrorResponse'
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/templates/rejected-document/preview'
        requestParameters:
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: POST

  #
  # Generate Rejected Signature Template Preview
  # --------------------------------------------------------------------------------
  '/templates/rejected-signature/preview':
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Generate Rejected Signature Template Preview
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    post:
      summary: Generates a text preview for a rejected signature template type and personalisation data
      description: Generates a text preview for a rejected signature template type and personalisation data
      operationId: generate-rejected-signature-template-preview
      tags:
        - Generate Rejected Signature Template Preview
      requestBody:
        $ref: '#/components/requestBodies/GenerateRejectedSignatureTemplatePreview'
      responses:
        '201':
          $ref: '#/components/responses/GenerateTemplatePreview'
        '400':
          $ref: '#/components/responses/400ErrorResponse'
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/templates/rejected-signature/preview'
        requestParameters:
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: POST

  #
  # Generate Requested Signature Template Preview
  # --------------------------------------------------------------------------------
  '/templates/requested-signature/preview':
    options:
      summary: CORS support
      description: |
        Enable CORS by returning correct headers
      tags:
        - Generate Requested Signature Template Preview
      responses:
        200:
          description: Default response for CORS method
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content: { }
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: |
            {
              "statusCode" : 200
            }
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
              method.response.header.Access-Control-Allow-Methods: '''*'''
              method.response.header.Access-Control-Allow-Origin: '''*'''
            responseTemplates:
              application/json: |
                {}
    post:
      summary: Generates a text preview for a requested signature template type and personalisation data
      description: Generates a text preview for a requested signature template type and personalisation data
      operationId: generate-requested-signature-template-preview
      tags:
        - Generate Requested Signature Template Preview
      requestBody:
        $ref: '#/components/requestBodies/GenerateRequestedSignatureTemplatePreview'
      responses:
        '201':
          $ref: '#/components/responses/GenerateTemplatePreview'
        '400':
          $ref: '#/components/responses/400ErrorResponse'
      security:
        - eroUserCognitoUserPoolAuthorizer: [ ]
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        uri: '${base_uri}/templates/requested-signature/preview'
        requestParameters:
        responseParameters:
          method.response.header.Access-Control-Allow-Headers: '''Content-Type,X-Amz-Date,Authorization,X-Api-Key'''
          method.response.header.Access-Control-Allow-Methods: '''*'''
          method.response.header.Access-Control-Allow-Origin: '''*'''
        connectionType: VPC_LINK
        connectionId: '${vpc_connection_id}'
        httpMethod: POST

  #
  # Generate Nino Not matched Template Preview
  # --------------------------------------------------------------------------------
...
components:
...
  responses:
    400ErrorResponse:
      description: Error response for a HTTP 400 describing fields in the request payload that are in error
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2022-09-28T18:01:42.105Z'
            status: 400
            error: 'Bad Request'
            message: 'Validation failed for object=applicationRequest. Error count: 14'
            validationErrors: 'Error on field applicant.nino: rejected value [aaaaaaaaaaa], must match ^.{1,10}$'
...
    GenerateTemplatePreview:
      description: Response containing the result of template previews in text and html format
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenerateTemplatePreviewResponse'
...
  requestBodies:
    GeneratePhotoResubmissionTemplatePreview:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GeneratePhotoResubmissionTemplatePreviewRequest'
    GenerateIdDocumentResubmissionTemplatePreview:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenerateIdDocumentResubmissionTemplatePreviewRequest'
    GenerateIdDocumentRequiredTemplatePreview:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenerateIdDocumentRequiredTemplatePreviewRequest'
    GenerateApplicationRejectedTemplatePreview:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenerateApplicationRejectedTemplatePreviewRequest'
...
    GenerateApplicationApprovedTemplatePreview:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenerateApplicationApprovedTemplatePreviewRequest'
    GenerateRejectedDocumentTemplatePreview:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenerateRejectedDocumentTemplatePreviewRequest'
...
    GenerateRejectedSignatureTemplatePreview:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenerateRejectedSignatureTemplatePreviewRequest'
    GenerateRequestedSignatureTemplatePreview:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenerateRequestedSignatureTemplatePreviewRequest'
...
  securitySchemes:
    eroUserCognitoUserPoolAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
      x-amazon-apigateway-authorizer:
        providerARNs:
          - '${cognito_ero_user_pool_arn}'
        type: cognito_user_pools
    awsIamAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: awsSigv4